task: refactor Step 2 — move auth web to user.infrastructure.web (Task 178)
Move AuthResource and its DTOs (LoginRequest, ActivateAccountRequest,
CurrentUserResponse, ActivationTokenResponse) plus the web-slice tests
(AuthResourceTest, ActivateAccountRequestTest) from iam.auth.infrastructure.web
to iam.user.infrastructure.web. Relocate the __files binding-error fixture
iam/auth/web → iam/user/web and repoint its path in AuthResourceTest.
Repoint the LoginRequest import in the two still-in-auth callers
(ActivateAccountIntegrationTest, AuthApi). Redundant package-info files
(main + test) dropped — target packages already provide them / follow the
no-test-package-info convention. URL paths unchanged.

Issue: #178
