+++ ProductSpecification/tasks/178-refactoring-collapse-iam-auth/progress.md
@@ -12,11 +12,11 @@ Type: refactoring
 - [x] refactor (cleanup: verify imports/static-analysis)
 
 ### Step 2: Move auth web auth.infrastructure.web → user.infrastructure.web
-- [~] refactor (move AuthResource + LoginRequest/ActivateAccountRequest/CurrentUserResponse/ActivationTokenResponse + AuthResourceTest/ActivateAccountRequestTest; move __files/iam/auth/web → __files/iam/user/web; update package + imports + resource paths)
-- [ ] refactor (cleanup: verify imports/static-analysis)
+- [x] refactor (move AuthResource + LoginRequest/ActivateAccountRequest/CurrentUserResponse/ActivationTokenResponse + AuthResourceTest/ActivateAccountRequestTest; move __files/iam/auth/web → __files/iam/user/web; update package + imports + resource paths)
+- [x] refactor (cleanup: verify imports/static-analysis)
 
 ### Step 3: Move auth integration tests + fixtures iam.auth → iam.user
-- [ ] refactor (move integration tests + fixtures package; update package + imports)
+- [~] refactor (move integration tests + fixtures package; update package + imports)
 - [ ] refactor (cleanup: verify imports/static-analysis)
 
 ### Step 4: Remove empty auth package and verify
