--- src/main/java/by/iivanov/rpm/iam/auth/infrastructure/web/package-info.java
@@ -1,6 +1,0 @@
-@NullMarked
-@InfrastructureRing
-package by.iivanov.rpm.iam.auth.infrastructure.web;
-
-import org.jmolecules.architecture.onion.simplified.InfrastructureRing;
-import org.jspecify.annotations.NullMarked;
