+++ src/test/java/by/iivanov/rpm/iam/auth/fixtures/AuthApi.java
@@ -1,6 +1,6 @@
 package by.iivanov.rpm.iam.auth.fixtures;
 
-import by.iivanov.rpm.iam.auth.infrastructure.web.LoginRequest;
+import by.iivanov.rpm.iam.user.infrastructure.web.LoginRequest;
 import by.iivanov.rpm.testing.api.AbstractApi;
 import by.iivanov.rpm.testing.api.AssertionResponse;
 import by.iivanov.rpm.testing.api.WebApi;
