--- src/test/java/by/iivanov/rpm/iam/auth/infrastructure/web/package-info.java
@@ -1,4 +1,0 @@
-@NullMarked
-package by.iivanov.rpm.iam.auth.infrastructure.web;
-
-import org.jspecify.annotations.NullMarked;
